requirements.txt
streamlit
pymupdf
reportlab
pytesseract
opencv-python
numpy
Pillow
